Initial commit
Setups some boilerplate for webpack, and react. Added in some bits to
support Barracks, and Sheet Router. Figuring out how to glue them
together based on looking at the choo.js code. Also, there is a bug with
webpack and/or chrome where you get an error about wrapOnError being
undefined in barracks.js. This was fixed by telling webpack devtool to
be source-map. This might be a bit slower, but at least hot reloading
and all that work now.